AB#32307 flatten AI prompt editing
