--- applications/Unity.GrantManager/modules/Unity.AI/src/Unity.AI.Web/Pages/Prompts/Versions/CreateVersionModal.cshtml.cs
@@ -1,34 +1,0 @@
-using Microsoft.AspNetCore.Mvc;
-using System;
-using System.Threading.Tasks;
-using Unity.AI.Prompts;
-using Volo.Abp.AspNetCore.Mvc.UI.RazorPages;
-
-namespace Unity.AI.Web.Pages.Prompts.Versions;
-
-public class CreateVersionModalModel : AbpPageModel
-{
-    [BindProperty]
-    public CreateUpdateAIPromptVersionDto Version { get; set; } = new();
-
-    private readonly IAIPromptVersionAppService _versionAppService;
-
-    public CreateVersionModalModel(IAIPromptVersionAppService versionAppService)
-    {
-        _versionAppService = versionAppService;
-    }
-
-    public void OnGet(Guid promptId)
-    {
-        Version = new CreateUpdateAIPromptVersionDto
-        {
-            PromptId = promptId
-        };
-    }
-
-    public async Task<IActionResult> OnPostAsync()
-    {
-        await _versionAppService.CreateAsync(Version);
-        return NoContent();
-    }
-}
